Add grade endpoint
