+++ backend/app/db/models/mixins/__init__.py
@@ -1,4 +1,5 @@
-from .mixin import UUIDMixin, Salary, Grade, Profession, Title, Note
+from .mixin import UUIDMixin, Salary, Profession, Title, Note
+from .grade import Grade
 
 __all__ = (
     "UUIDMixin",
