+++ backend/app/db/models/mixins/grade.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+
+class Grade(Enum):
+    junior = "junior"
+    middle = "middle"
+    senior = "senior"
+    lead = "lead"
+    principal = "principal"
+
+    @classmethod
+    def __iter__(cls):
+        return iter(cls._member_names_)
